somewhat cleanup warnings

+++ ElDorito/Source/Blam/Tags/Sounds/Sound.hpp
@@ -210,7 +210,7 @@ namespace Blam::Tags::Sounds
 			Math::Bounds<int16_t> Bend;
 			Math::Bounds<int16_t> MaxGainPitch;
 			Math::Bounds<int16_t> Unknown2;
-			uint32_t Unknown3;
+			float Unknown3;
 			uint32_t Unknown4;
 			uint32_t Unknown5;
 			uint32_t Unknown6;
